web-app/backend/app/requirements.txt
ensemble_boxes
fastapi
imageio
ipython
keras
loguru
matplotlib
numpy
opencv_python
pandas
Pillow
pycocotools
pydantic
PyYAML
scipy
segmentation_models
six
scikit-image
tensorflow
tifffile
tqdm
uvicorn